🔥 chore: refactor project structure
